requirements.txt
beautifulsoup4==4.12.2
Jinja2==3.1.2
locustio==0.999
Requests==2.31.0
selenium==4.11.2
